refactor: items  organization folder
