+++ src/app/modules/items/mocks/item-mock.model.ts
@@ -5,5 +5,6 @@ export const itemMockModel: ItemModel = {
   description: 'Cascos usados. No son muy buenos, pero son bastante bonitos. Precio negociable',
   price: '20',
   email: 'headphones@wallapop.com',
-  image: 'https://frontend-tech-test-data.s3-eu-west-1.amazonaws.com/img/headphones.jpg'
+  image: 'https://frontend-tech-test-data.s3-eu-west-1.amazonaws.com/img/headphones.jpg',
+  favorite: false
 };
