+++ src/app/modules/items/models/item.model.ts
@@ -1,0 +1,8 @@
+import { ProductModel } from '../../../core/models/product.model';
+
+export class ItemModel extends ProductModel {
+  description: string = '';
+  price: string = '';
+  email: string = '';
+  image: string = '';
+}
