+++ src/app/modules/items/services/items.facade.spec.ts
@@ -4,8 +4,8 @@ import { of } from 'rxjs';
 import { ItemsFacade } from './items.facade';
 import { IQueryParams } from '../../api/interfaces/pagination.interface';
 import { ItemsService } from './items/items.service';
-import { itemMockModel } from './items/mocks/item-mock.model';
-import { ItemModel } from './items/models/item.model';
+import { itemMockModel } from '../mocks/item-mock.model';
+import { ItemModel } from '../models/item.model';
 import { ItemsState } from '../state/items.reducer';
 import { Store, StoreModule } from '@ngrx/store';
 import * as fromItems from '../state/items.reducer';
