+++ src/app/modules/items/services/items.facade.ts
@@ -3,7 +3,7 @@ import { select, Store } from '@ngrx/store';
 import { IQueryParams } from '../../api/interfaces/pagination.interface';
 import { getItems, resetStateItems } from '../state/items.actions';
 import { ItemsState } from '../state/items.reducer';
-import { ItemModel } from './items/models/item.model';
+import { ItemModel } from '../models/item.model';
 import { selectItems, selectTotal, selectLoading } from '../state/items.selector';
 import { Observable } from 'rxjs';
 import { AppFacade } from '../../../core/services/app.facade';
