+++ src/app/modules/items/services/items/items.service.spec.ts
@@ -10,9 +10,9 @@ import {
 } from '@api';
 import { of } from 'rxjs';
 import { ItemsService } from './items.service';
-import { itemMockModel } from './mocks/item-mock.model';
-import { ItemModel } from './models/item.model';
-import { ItemTranslator } from './translate/item.translator';
+import { itemMockModel } from '../../mocks/item-mock.model';
+import { ItemModel } from '../../models/item.model';
+import { ItemTranslator } from '../../translate/item.translator';
 
 const itemContracts: ItemContract[] = [
   itemMockContract,
