--- src/app/modules/items/services/items/models/item.model.ts
@@ -1,8 +1,0 @@
-
-export class ItemModel {
-  title: string = '';
-  description: string = '';
-  price: string = '';
-  email: string = '';
-  image: string = '';
-}
