+++ src/app/modules/items/state/items.actions.ts
@@ -1,6 +1,6 @@
 import { createAction, props } from '@ngrx/store';
 import { HttpErrorResponse } from '@angular/common/http';
-import { ItemModel } from '../services/items/models/item.model';
+import { ItemModel } from '../models/item.model';
 import { IQueryParams } from '../../api/interfaces/pagination.interface';
 import { IApiResponse } from '../../api/interfaces/response.interface';
