+++ src/app/modules/items/translate/item.translator.ts
@@ -1,11 +1,11 @@
 import { ItemModel } from '../models/item.model';
-import { ItemContract } from '../../../../api/services/api-items/contracts/item.contract';
+import { ItemContract } from '../../api/services/api-items/contracts/item.contract';
 
 
 export class ItemTranslator {
 
   static translateContractToModel(contract: ItemContract): ItemModel {
-    const model: ItemModel = { ...contract };
+    const model: ItemModel = { ...contract, favorite: false };
     return model;
   }
